fix: keep fractional SV2 pool difficulty to avoid difficulty-too-low rejects
sv2_target_to_pdiff truncated the difficulty to an integer, so a fractional target (e.g. 931.1) became 931. The submit threshold was then too low and shares in [931, 931.1) were sent and rejected by the pool as difficulty-too-low. Return a double to keep full precision.

+++ components/stratum_v2/include/sv2_protocol.h
@@ -197,7 +197,9 @@ void sv2_ext_job_free(sv2_ext_job_t *job);
 
 // --- Helpers ---
 
-// Convert U256 LE target to pool difficulty (pdiff)
-uint32_t sv2_target_to_pdiff(const uint8_t target[32]);
+// Convert U256 LE target to pool difficulty (pdiff). Returns a double to
+// preserve fractional difficulty (e.g. 931.1) so the submit threshold matches
+// the pool's exact target.
+double sv2_target_to_pdiff(const uint8_t target[32]);
 
 #endif /* SV2_PROTOCOL_H */
